# Avoid using union structs for asfloat(int) and friends