--- a/Unity.Mathematics.TestProject/Packages/manifest.json
+++ b/Unity.Mathematics.TestProject/Packages/manifest.json
@@ -1,5 +1,6 @@
 {
   "dependencies": {
+    "com.s-schoener.asmexplorer": "file:/Users/dale/code/unity-asm-explorer-package/Packages/com.s-schoener.asmexplorer",
     "com.unity.burst": "1.2.3",
     "com.unity.ide.rider": "1.1.4",
     "com.unity.ide.vscode": "1.1.4",