… src/CHANGELOG.md
… src/Unity.Mathematics.PerformanceTests/TestConversions.gen.cs
… src/Unity.Mathematics.PerformanceTests/TestHash.gen.cs
… src/Unity.Mathematics.PerformanceTests/TestInverse.gen.cs
… src/Unity.Mathematics.PerformanceTests/TestMath.gen.cs
… src/Unity.Mathematics.PerformanceTests/TestMinMaxAABB.gen.cs
… src/Unity.Mathematics.PerformanceTests/TestMul.gen.cs
… src/Unity.Mathematics.PerformanceTests/TestNoise.gen.cs
… src/Unity.Mathematics.PerformanceTests/TestNormalize.gen.cs
… src/Unity.Mathematics.PerformanceTests/TestPlane.gen.cs
… src/Unity.Mathematics.PerformanceTests/TestRandom.gen.cs
… src/Unity.Mathematics.PerformanceTests/TestRotation.gen.cs
… src/Unity.Mathematics.PerformanceTests/TestShuffle.gen.cs
… src/Unity.Mathematics.PerformanceTests/TestTranspose.gen.cs
… src/Unity.Mathematics.PerformanceTests/TestTrig.gen.cs
… src/Unity.Mathematics/math.cs