@sschoener 2021-11-24
Nice! Can you show the ASM difference between the old and new `asuint(int)`? You'd expect a cast to be free, so I'm curious what it did before and what it does now.
I'm curious as to whether you checked the codegen for Burst as well? It's not clear to me what N you are using for your perf tests and whether some of the perf degradations for Burst are noise or actual real things (for which there would be a codegen explanation).
My first thought was "we should have a define for whether this is compiled with Burst!" but then I realized that this is not how Burst works _at all_, so we'll have to settle for a compromise between Mono and Burst (and that compromise better doesn't hit Mono as bad as our previous compromise).