@unpacklo 2021-11-25
> Nice! Can you show the ASM difference between the old and new `asuint(int)`?

Old:

```asm
public static uint asuint(int x)
Address: 000000017CE19200
Code Size in Bytes: 16
Debug mode: enabled

000000017ce19200 48 83 ec 08                    sub rsp, 0x8
000000017ce19204 48 89 3c 24                    mov [rsp], rdi
000000017ce19208 48 8b c7                       mov rax, rdi
000000017ce1920b 48 83 c4 08                    add rsp, 0x8                 ; 8
000000017ce1920f c3                             ret
```

New:

```asm
public static uint asuint(int x)
Address: 000000017CA2CA80
Code Size in Bytes: 16
Debug mode: enabled

000000017ca2ca80 48 83 ec 08                    sub rsp, 0x8
000000017ca2ca84 48 89 3c 24                    mov [rsp], rdi
000000017ca2ca88 8b 04 24                       mov eax, [rsp]               ; read singlestep trampoline
000000017ca2ca8b 48 83 c4 08                    add rsp, 0x8                 ; 8
000000017ca2ca8f c3                             ret
```

I don't quite understand why casting vs type punning makes any difference here, but I also just realized I don't have perf tests for these so I'll add those and see how those perform for this case.

> I'm curious as to whether you checked the codegen for Burst as well?

Not for all, but some. For example, here's `asfloat(uint3)`.

Old:

```asm
        .text
        .intel_syntax noprefix
        .file        "main"
        .globl        "Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.BurstTestFunction(ref Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.Arguments args)_5B581B7AAD6CC3EF" # -- Begin function Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.BurstTestFunction(ref Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.Arguments args)_5B581B7AAD6CC3EF
        .p2align        4, 0x90
        .type        "Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.BurstTestFunction(ref Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.Arguments args)_5B581B7AAD6CC3EF",@function
"Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.BurstTestFunction(ref Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.Arguments args)_5B581B7AAD6CC3EF": # @"Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.BurstTestFunction(ref Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.Arguments args)_5B581B7AAD6CC3EF"
.Lfunc_begin0:
        .file        1 "/Users/dale/code/Unity.Mathematics/src/Unity.Mathematics.PerformanceTests/TestMath.gen.cs"
        .loc        1 1192 0                # TestMath.gen.cs:1192:0
        .cfi_sections .debug_frame
        .cfi_startproc
# %bb.0:                                # %entry
        mov        rax, -4800000
        .p2align        4, 0x90
.LBB0_1:                                # %BL.0004.i
                                        # =>This Inner Loop Header: Depth=1
.Ltmp0:
        .loc        1 1178 49 prologue_end  # TestMath.gen.cs:1178:49
        mov        rcx, qword ptr [rdi + 8]
        mov        rdx, qword ptr [rdi + 16]
        movss        xmm0, dword ptr [rdx + rax + 4800008] # xmm0 = mem[0],zero,zero,zero
        movsd        xmm1, qword ptr [rdx + rax + 4800000] # xmm1 = mem[0],zero
        movss        dword ptr [rcx + rax + 4800000], xmm1
        extractps        dword ptr [rcx + rax + 4800004], xmm1, 1
        movss        dword ptr [rcx + rax + 4800008], xmm0
        .loc        1 1182 13               # TestMath.gen.cs:1182:13
        add        rax, 12
        jne        .LBB0_1
.Ltmp1:
# %bb.2:                                # %"Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.CommonTestFunction(ref Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.Arguments args)_A3E8969154AF482D.exit"
        .loc        1 1193 13               # TestMath.gen.cs:1193:13
        ret
```

New:

```asm
        .text
        .intel_syntax noprefix
        .file        "main"
        .globl        "Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.BurstTestFunction(ref Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.Arguments args)_5B581B7AAD6CC3EF" # -- Begin function Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.BurstTestFunction(ref Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.Arguments args)_5B581B7AAD6CC3EF
        .p2align        4, 0x90
        .type        "Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.BurstTestFunction(ref Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.Arguments args)_5B581B7AAD6CC3EF",@function
"Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.BurstTestFunction(ref Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.Arguments args)_5B581B7AAD6CC3EF": # @"Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.BurstTestFunction(ref Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.Arguments args)_5B581B7AAD6CC3EF"
.Lfunc_begin0:
        .file        1 "/Users/dale/code/Unity.Mathematics/src/Unity.Mathematics.PerformanceTests/TestMath.gen.cs"
        .loc        1 1192 0                # TestMath.gen.cs:1192:0
        .cfi_sections .debug_frame
        .cfi_startproc
# %bb.0:                                # %entry
        mov        rax, -4800000
        .p2align        4, 0x90
.LBB0_1:                                # %BL.0004.i
                                        # =>This Inner Loop Header: Depth=1
.Ltmp0:
        .loc        1 1178 49 prologue_end  # TestMath.gen.cs:1178:49
        mov        rcx, qword ptr [rdi + 8]
        mov        rdx, qword ptr [rdi + 16]
        movss        xmm0, dword ptr [rdx + rax + 4800008] # xmm0 = mem[0],zero,zero,zero
        movsd        xmm1, qword ptr [rdx + rax + 4800000] # xmm1 = mem[0],zero
        movss        dword ptr [rcx + rax + 4800000], xmm1
        extractps        dword ptr [rcx + rax + 4800004], xmm1, 1
        movss        dword ptr [rcx + rax + 4800008], xmm0
        .loc        1 1182 13               # TestMath.gen.cs:1182:13
        add        rax, 12
        jne        .LBB0_1
.Ltmp1:
# %bb.2:                                # %"Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.CommonTestFunction(ref Unity.Mathematics.PerformanceTests.TestMath.asfloat_uint3.Arguments args)_A3E8969154AF482D.exit"
        .loc        1 1193 13               # TestMath.gen.cs:1193:13
        ret
```

The code is identical in this case which is good but I don't know if this is true for all the code that got changed. I'm going to grab all the bursted code and then see if a diff surfaces anything of interest.