@unpacklo 2021-12-02
@sschoener I've compared all the burst code gen and there are no diffs with this code change. Regarding the mono generated code for type punning vs casting in the case of `asuint(int)`, I haven't been able to determine whether one version is in fact faster than the other on my MacBook Pro, but it seems pretty safe to assume that accessing the stack can only be in the best case as fast as the register move.